Refactor css classes + markdown highlighting for articles
Remove "j-" namespaced classes; these are unnecessary
